move fertilizer factory into core
